+++ src/OpenTelemetry.AutoInstrumentation.Native/assembly_redirection_net.h
@@ -7,7 +7,7 @@
 
 #define ASSEMBLY_REDIRECTION_NET \
     { 80, { \
-        { _W("Google.Protobuf"), {3, 36, 1, 0} }, \
+        { _W("Google.Protobuf"), {3, 36, 2, 0} }, \
         { _W("Microsoft.Extensions.Configuration"), {8, 0, 0, 0} }, \
         { _W("Microsoft.Extensions.Configuration.Abstractions"), {8, 0, 0, 0} }, \
         { _W("Microsoft.Extensions.Configuration.Binder"), {8, 0, 0, 0} }, \
@@ -44,7 +44,7 @@
         { _W("OpenTelemetry.Instrumentation.SqlClient"), {1, 18, 0, 1309} }, \
         { _W("OpenTelemetry.Instrumentation.StackExchangeRedis"), {1, 18, 0, 1354} }, \
         { _W("OpenTelemetry.Instrumentation.Wcf"), {1, 18, 0, 1310} }, \
-        { _W("OpenTelemetry.OpAmp.Client"), {0, 7, 0, 1339} }, \
+        { _W("OpenTelemetry.OpAmp.Client"), {0, 7, 0, 1355} }, \
         { _W("OpenTelemetry.Resources.Azure"), {1, 18, 0, 1345} }, \
         { _W("OpenTelemetry.Resources.Container"), {1, 18, 0, 1316} }, \
         { _W("OpenTelemetry.Resources.Host"), {1, 18, 0, 1319} }, \
@@ -60,7 +60,7 @@
         { _W("System.ServiceModel.Primitives"), {4, 7, 0, 0} }, \
     }}, \
     { 90, { \
-        { _W("Google.Protobuf"), {3, 36, 1, 0} }, \
+        { _W("Google.Protobuf"), {3, 36, 2, 0} }, \
         { _W("Microsoft.Extensions.Configuration"), {9, 0, 0, 0} }, \
         { _W("Microsoft.Extensions.Configuration.Abstractions"), {9, 0, 0, 0} }, \
         { _W("Microsoft.Extensions.Configuration.Binder"), {9, 0, 0, 0} }, \
@@ -97,7 +97,7 @@
         { _W("OpenTelemetry.Instrumentation.SqlClient"), {1, 18, 0, 1309} }, \
         { _W("OpenTelemetry.Instrumentation.StackExchangeRedis"), {1, 18, 0, 1354} }, \
         { _W("OpenTelemetry.Instrumentation.Wcf"), {1, 18, 0, 1310} }, \
-        { _W("OpenTelemetry.OpAmp.Client"), {0, 7, 0, 1339} }, \
+        { _W("OpenTelemetry.OpAmp.Client"), {0, 7, 0, 1355} }, \
         { _W("OpenTelemetry.Resources.Azure"), {1, 18, 0, 1345} }, \
         { _W("OpenTelemetry.Resources.Container"), {1, 18, 0, 1316} }, \
         { _W("OpenTelemetry.Resources.Host"), {1, 18, 0, 1319} }, \
@@ -113,7 +113,7 @@
         { _W("System.ServiceModel.Primitives"), {4, 7, 0, 0} }, \
     }}, \
     { 100, { \
-        { _W("Google.Protobuf"), {3, 36, 1, 0} }, \
+        { _W("Google.Protobuf"), {3, 36, 2, 0} }, \
         { _W("Microsoft.Extensions.Configuration"), {10, 0, 0, 0} }, \
         { _W("Microsoft.Extensions.Configuration.Abstractions"), {10, 0, 0, 0} }, \
         { _W("Microsoft.Extensions.Configuration.Binder"), {10, 0, 0, 0} }, \
@@ -150,7 +150,7 @@
         { _W("OpenTelemetry.Instrumentation.SqlClient"), {1, 18, 0, 1309} }, \
         { _W("OpenTelemetry.Instrumentation.StackExchangeRedis"), {1, 18, 0, 1354} }, \
         { _W("OpenTelemetry.Instrumentation.Wcf"), {1, 18, 0, 1310} }, \
-        { _W("OpenTelemetry.OpAmp.Client"), {0, 7, 0, 1339} }, \
+        { _W("OpenTelemetry.OpAmp.Client"), {0, 7, 0, 1355} }, \
         { _W("OpenTelemetry.Resources.Azure"), {1, 18, 0, 1345} }, \
         { _W("OpenTelemetry.Resources.Container"), {1, 18, 0, 1316} }, \
         { _W("OpenTelemetry.Resources.Host"), {1, 18, 0, 1319} }, \